extracted common functionality out of GeneratorStatement into InterpolatedStatement
