+++ src/game/gdl/statement/StatementType.groovy
@@ -1,0 +1,13 @@
+package game.gdl.statement
+
+/**
+ * @author Lawrence Thatcher
+ *
+ * An enumeration of the potential types of statements
+ */
+enum StatementType
+{
+	Simple,
+	Generator,
+	Substitution
+}
