+++ test/game/constructs/board/TestBoard.groovy
@@ -25,7 +25,7 @@ class TestBoard
 		// parametrized
 		gdl = board.getImplementation(GameFunction.N_inARow(3))
 		assert gdl instanceof DynamicComponentsClause
-		assert gdl.contains("(<= (3inARow ?w) (row ?x ?y ?w))")
+		assert gdl.contains("(<= (3inARow ?w) (row3 ?x ?y ?w))")
 		assert gdl.statements.size() == 4
 
 	}
